+++ packages/runtime-core/src/index.ts
@@ -1,4 +1,4 @@
 export { onBeforeMount, onBeforeUnMount, onBeforeUpdate, onMounted, onUnMounted, onUpdated } from './life-cycles'
-export { $computed, $reactive, $ref, computed, effect, manualTrack, reactive, ref } from './reactive'
+export { $computed, $reactive, $ref, computed, effect, manualTrack, nv, reactive, ref } from './reactive'
 export { Router } from './router'
 export { h, mount } from './vnode'